refactor: rename module from obsidian to plugin
