+++ src/plugin/commands/index.ts
