+++ src/plugin/commands/show-date-chooser.ts
@@ -1,5 +1,5 @@
 import type { Editor } from 'obsidian';
-import type { ReminderPluginUI } from 'obsidian/ui';
+import type { ReminderPluginUI } from 'plugin/ui';
 
 export function showDateChooser(checking: boolean, editor: Editor, ui: ReminderPluginUI): boolean | void {
   if (checking) {
