+++ src/plugin/commands/show-reminder-list.ts
@@ -1,4 +1,4 @@
-import type { ReminderPluginUI } from 'obsidian/ui';
+import type { ReminderPluginUI } from 'plugin/ui';
 
 export function showReminderList(checking: boolean, ui: ReminderPluginUI) {
   if (!checking) {
