+++ src/plugin/commands/toggle-checklist-status.ts
