+++ src/plugin/obsidian-hack/command.ts
