+++ src/plugin/obsidian-hack/index.ts
