+++ src/plugin/obsidian-hack/plugin.ts
