+++ src/plugin/settings.ts
