+++ src/plugin/ui/date-chooser-modal.ts
