+++ src/plugin/vault.ts
